% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.irf.R
\name{extract.irf}
\alias{extract.irf}
\title{Extract impulse-response series from irf objects}
\usage{
extract.irf(object, impulse, response, probs = 0.5)
}
\arguments{
\item{object}{A irf object.}

\item{impulse}{Impulse variable name or index.}

\item{response}{Response Variable name or index.}

\item{probs}{numeric vector of probabilities with values in [0,1].}
}
\description{
Extracts the desired impulse-response serie from irf objects for a given quantile.
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##

# = This is an ilustrative example = #
# = The identification ignores which variables are more exogenous = #
data("BRinf")
Y=BRinf[,1:59]# remove expectation variables
modelB=lbvar(Y,p=3)
identB=identification(modelB)
irfB=irf(modelB,identB,h=12,boot = TRUE,M=100)
extract.irf(irfB,1,2,probs=0.5)


}
\references{
Garcia, Medeiros and Vasconcelos (2017).
}
\seealso{
\code{\link{predict}}, \code{\link{lbvar}}, \code{\link{identification}}, \code{\link{irf}}
}
\keyword{Bayesian}
\keyword{High-dimension,}
\keyword{VAR,}
\keyword{irf,}
\keyword{models}
