% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lbvar.R
\name{predict.lbvar}
\alias{predict.lbvar}
\alias{coef.lbvar}
\title{Functions for lbvar objects}
\usage{
\method{predict}{lbvar}(object, h = 0, newdata = NULL, ...)

\method{coef}{lbvar}(object, type = "equation", ...)
}
\arguments{
\item{object}{lbvar object generated by the function lbvar.}

\item{h}{Forecasting horizon. If h=0 returns the fitted values.}

\item{newdata}{Exogenous variables to compute the predictions. If the model has exogenous controls (xreg) newdata must be supplied.}

\item{...}{Arguments for other methods.}

\item{type}{equation to extract coefficients by equation and block to extract blocks like cons. lags and xreg coefficients.}
}
\description{
Functions for lbvar objects
}
\examples{
## == This example uses the Brazilian inflation data from
#Garcia, Medeiros and Vasconcelos (2017) == ##
data("BRinf")
Y=BRinf[,1:59]# remove expectation variables

modelB=lbvar(Y,p=4)
predict(modelB,h=10)

# take a look at the coefficients
eq=coef(modelB,type="equation")
block=coef(modelB,type="block")
block$Lag1

}
\references{
Garcia, Medeiros and Vasconcelos (2017).
}
\seealso{
\code{\link{lbvar}}
}
